- replace the assignability cache and HierarchyIndexer with AssignabilityCheck and check supertypes instead of subtypes; this works better with computing index and allows the check to cache the misses
- match extension method name first (before checking the assignability of the first parameter)
- dev mode - do not regenerate the value resolvers for non-application classes during hot reload

I can observe ~50% speedup of the `QuteProcessor#validateExpressions` build step (100ms -> 45ms in prod build, 250ms -> 140ms in the dev mode) in a test app that has roughly 400 qute output expressions to validate.