@gastaldi 2023-04-16
Backport to 3.0?